perf: 完善 tiktok 爬虫功能

+++ backend-src/Plugins/UzonMailDB/SQL/EmailCrawler/TikTokAuthor.cs
@@ -32,13 +32,13 @@ namespace UZonMail.DB.SQL.EmailCrawler
         /**
          * 统计相关信息
          */
-        public int DiggCount { get; set; }
-        public int FollwerCount { get; set; }
-        public int FollwingCount { get; set; }
-        public int FreindCount { get; set; }
-        public int Heart { get; set; }
-        public int HeartCount { get; set; }
-        public int VideoCount { get; set; }
+        public long DiggCount { get; set; }
+        public long FollwerCount { get; set; }
+        public long FollwingCount { get; set; }
+        public long FreindCount { get; set; }
+        public long Heart { get; set; }
+        public long HeartCount { get; set; }
+        public long VideoCount { get; set; }
 
         /**
          * 解析后的数据
